requirements.txt
blinker==1.9.0
click==8.3.3
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
Werkzeug==3.1.8
